feat: authorization API

+++ src/pages/api/[[...user]].ts
@@ -1,5 +1,6 @@
 import { deleteData, retrieveData, updateData } from "@/lib/firebase/service";
 import type { NextApiRequest, NextApiResponse } from "next";
+import jwt from "jsonwebtoken";
 
 export default async function handler(
   req: NextApiRequest,
@@ -15,31 +16,64 @@ export default async function handler(
       .status(200)
       .json({ status: true, statusCode: 200, message: "success", data });
   } else if (req.method === "PUT") {
-    const { id, data } = req.body;
-    await updateData("users", id, data, (result: boolean) => {
-      if (result) {
-        res
-          .status(200)
-          .json({ status: true, statusCode: 200, message: "success" });
-      } else {
-        res
-          .status(400)
-          .json({ status: false, statusCode: 400, message: "failed" });
+    const { data } = req.body;
+    const { user }: any = req.query;
+    const token = req.headers.authorization?.split(" ")[1] || "";
+    console.log("user: ", user);
+
+    jwt.verify(
+      token,
+      process.env.NEXTAUTH_SECRET || "",
+      async (err: any, decoded: any) => {
+        console.log("decoded: ", decoded);
+        if (decoded && decoded.role === "admin") {
+          await updateData("users", user[1], data, (result: boolean) => {
+            if (result) {
+              res
+                .status(200)
+                .json({ status: true, statusCode: 200, message: "success" });
+            } else {
+              res
+                .status(400)
+                .json({ status: false, statusCode: 400, message: "failed" });
+            }
+          });
+        } else {
+          console.log(err);
+          res
+            .status(403)
+            .json({ status: false, statusCode: 403, message: "Access denied" });
+        }
       }
-    });
+    );
   } else if (req.method === "DELETE") {
     const { user }: any = req.query;
-    console.log("userId: ", user);
-    await deleteData("users", user[1], (result: boolean) => {
-      if (result) {
-        res
-          .status(200)
-          .json({ status: true, statusCode: 200, message: "success" });
-      } else {
-        res
-          .status(400)
-          .json({ status: false, statusCode: 400, message: "failed" });
+    const token = req.headers.authorization?.split(" ")[1] || "";
+
+    jwt.verify(
+      token,
+      process.env.NEXTAUTH_SECRET || "",
+      async (err: any, decoded: any) => {
+        console.log("decoded: ", decoded);
+        if (decoded && decoded.role === "admin") {
+          await deleteData("users", user[1], (result: boolean) => {
+            if (result) {
+              res
+                .status(200)
+                .json({ status: true, statusCode: 200, message: "success" });
+            } else {
+              res
+                .status(400)
+                .json({ status: false, statusCode: 400, message: "failed" });
+            }
+          });
+        } else {
+          console.log(err);
+          res
+            .status(403)
+            .json({ status: false, statusCode: 403, message: "Access denied" });
+        }
       }
-    });
+    );
   }
 }
